requirements.txt
requests
pyexiftool
